Chapter 1-3 Project
Create a program using IntelliJ to get and display basic information about 3 pairs of shoes.

There will be 2 classes in the program - a Main class, and a PairOfShoes class.

The PairOfShoes class will define private fields for the brand name of the shoes, the size of the shoes, and the color of the shoes, and public methods for getting and setting the value of the fields, as well as a public method (printAll()) for printing (without a newline character) the following:
"colorOfShoes brandName size sizeOfShoes"
(ex: "White Nike size 10.0")

The Main class will create 3 PairOfShoes objects. It will then prompt for and get the brand name, color and size of the shoes and store those items in the properties of the 3 PairOfShoes objects. Then it will display the color, brand name and size of each of the pair of shoes (using the printAll() method of the PairOfShoes class). Finally, using the getter methods, it will display the brand name of the 1st shoe object, the color of the 2nd shoe object, and the size of the 3rd shoe object (see the Example Run).

Both classes should have a main comments section at the top of the class with your name, date, and the name of the program. In addition, the Main class must also have a program description consisting of at least 2 sentences. Throughout both classes there should also be comments explaining what the various lines of code are doing.